moved an external module into an existing file for readability.
